requirements.txt
# Core Web Frameworks
django==4.1.3
djangorestframework==3.12.4
flask==2.3.3
werkzeug==2.3.7

# Database
pymongo==4.6.2
sqlalchemy==2.0.21

# HTTP and Requests
requests==2.27.0
httpx==0.24.1
aiohttp==3.8.5

# Data Science & Analysis
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scipy==1.11.2

# Testing
pytest==7.4.2
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-django==4.5.2

# Code Quality
black==23.7.0
flake8==6.1.0
pylint==2.17.5
mypy==1.5.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.3.0
celery==5.3.1
redis==4.6.0

# CLI
click==8.1.7
rich==13.5.2

# Date/Time
python-dateutil==2.8.2
pytz==2023.3
